backend/petsphere/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environs==11.2.0
marshmallow==3.23.1
packaging==24.2
pillow==11.0.0
psycopg-binary==3.2.3
psycopg2-binary==2.9.10
PyJWT==2.10.0
python-dotenv==1.0.1
sqlparse==0.5.2
typing_extensions==4.12.2
